requirements.txt
# Data
datasets>=2.19.0

# NLP / Embeddings
scikit-learn>=1.4.2
nltk>=3.8.1


# Classifiers
lightgbm>=4.3.0
xgboost>=2.0.0
scipy>=1.13.0

# Tracking
mlflow>=3.11.1

# Explainability
lime>=0.2.0.1

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.1
python-dotenv>=1.0.1

# Serialization
joblib>=1.4.0

# Dev / testing
pytest>=8.2.0
httpx>=0.27.0

# System monitoring
psutil>=5.9.0
